requirements.txt
fastapi
uvicorn
gunicorn
google-generativeai
opencv-python-headless
requests
python-multipart
python-dotenv
